feat: change all price field from float to int in order, order detail, product, and shipping rate entities
- Update entities: Order, OrderDetail, Product, and ShippingRate entities to use int instead of float for cost-related fields.
- Modify corresponding DTOs and mappers to reflect the data type changes.
- Adjust calculations in services to ensure compatibility with the new data types.

+++ shopme-common/src/main/java/com/shopme/common/entity/Order.java
@@ -51,11 +51,11 @@ public class Order {
 	@Column(name = "order_time")
 	private Date orderTime;
 	
-	private float shippingCost;
-	private float productCost;
-	private float subtotal;
-	private float tax;
-	private float total;
+	private int shippingCost;
+	private int productCost;
+	private int subtotal;
+	private int tax;
+	private int total;
 	
 	private int deliverDays;
 	private Date deliverDate;
