+++ shopme-webparent/shopme-client/src/main/java/com/shopme/client/dto/response/OrderItemResponse.java
@@ -11,5 +11,5 @@ public class OrderItemResponse {
     private String productName;
     private String productMainImage;
     private int quantity;
-    private float unitPrice;
+    private int unitPrice;
 }
